Use a Course object as a parameter for the upper TableView in TeacherDashboard + Clean up [Discovered BUGs]
Use Course object as a parameter (instead of int courseId) for the methods related to the TableView in Teacher Dashboard (upper right TableView).

Removed methods using mock classes (from Sprint 1). The methods are not used anymore.

Removed the unused AttendanceModel.

Discovered two problems:
(1) The today page is not loading.
(2) There is possibly a problem with the ComboBox binding for TeacherDashboard.
Check Output for details.